+++ src/index.ts
@@ -1,6 +1,6 @@
 export {World} from './world';
 export {Type} from './type';
-export {Entity} from './entity';
+export type {Entity} from './entity';
 export {System} from './system';
 export type {SystemType} from './system';
 export {component, field, system} from './decorators';