+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.8.2
+- Added `Entity.hold` and `Entity.isSame` methods.
+- Stopped returning `this` from `Entity.add` and `Entity.addAll`.  This is technically a breaking change and ought to be a major version bump, but it feels so trivial I didn't want to bother.
+- Started checking that `Entity` objects obtained from queries are not used for too long.  This reduced create/delete entity performance by 50% in dev mode, but didn't affect perf mode (where the check is bypassed).
+- Documented `Entity` class.
+
 ### 0.8.1
 - Improved query performance as the number of components grows.  This actually reduced performance on some of the synthetic benchmarks but should have a greatly positive effect in actual applications.
 - Improved performance of component initialization, both with and without custom field values.